pyproject.toml
[tool.poetry]
name = "scrat"
version = "0.1.0"
description = ""
authors = ["javiber"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.18"
appdirs = "^1.4.4"
click = "^8.1.4"
pyyaml = "^6.0"
rich = "^13.4.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
isort = "^5.12.0"
ipython = "^8.14.0"
pytest-mock = "^3.11.1"
ipdb = "^0.13.13"

[tool.poetry.scripts]
scrat = "scrat.cli:scrat"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
